+++ test/controllers/admin/broadcast_comments_controller_test.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+# Tests to assure that admins can review blog comments
+class Admin::BroadcastCommentsControllerTest < ActionController::TestCase
+  test 'should get index' do
+    login(users(:owner))
+    get :index
+    assert_response :success
+  end
+end